Ensure that pull service runs regardless if playbook changed
Previous behavior was to skip running the playbook if the repo had not
been updated since the last run. However, there are some tasks in here
that will check Github for new releases and deploy them every time they
are run, and of course those tasks were not running and keeping their
components up-to-date. We ensure the playbook runs every time to
accomodate these tasks.