Removed ff nightly tasks.
I don't know how to make it work without pulling in a role from the
internet, which wouldn't work with the ansible-pull because of course
that cannot really depend on anything that's not built-in.

Besides, I still don't know if the flatpak method is the best way on
Fedora, and I have no idea how to make things work on Ubuntu. It was a
half-baked module anyway.